+++ .circleci/config.yml
@@ -32,7 +32,6 @@ commands:
             sudo apt-get update
             sudo apt-get install -y --force-yes gconf-service libasound2 libatk1.0-0 libc6 libcairo2 libcups2 libdbus-1-3 libexpat1 libfontconfig1 libgcc1 libgconf-2-4 libgdk-pixbuf2.0-0 libglib2.0-0 libgtk-3-0 libnspr4 libpango-1.0-0 libpangocairo-1.0-0 libstdc++6 libx11-6 libx11-xcb1 libxcb1 libxcomposite1 libxcursor1 libxdamage1 libxext6 libxfixes3 libxi6 libxrandr2 libxrender1 libxss1 libxtst6 ca-certificates fonts-liberation libappindicator1 libnss3 lsb-release xdg-utils wget
 
-
 jobs:
   checkout:
     <<: *defaults
@@ -159,12 +158,6 @@ jobs:
           name: '`yarn docs:api` changes committed?'
           command: git diff --exit-code
       - run:
-          name: Can we generate the css-utils docs?
-          command: cd packages/material-ui-css-utils && yarn run docs
-      - run:
-          name: 'css-utils `yarn run docs` changes committed?'
-          command: git diff --exit-code packages/material-ui-css-utils
-      - run:
           name: Can we generate the @material-ui/core build?
           command: cd packages/material-ui && yarn build
       - run:
@@ -179,9 +172,6 @@ jobs:
       - run:
           name: Can we generate the @material-ui/system build?
           command: cd packages/material-ui-system && yarn build
-      - run:
-          name: Can we generate the @material-ui/css-utils build?
-          command: cd packages/material-ui-css-utils && yarn build
       - run:
           name: Prepare persisting workspace
           command: tar cfvz builds.tar.gz packages/*/build
