+++ docs/src/pages/customization/themes/ResponsiveFontSizes.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { createMuiTheme, responsiveFontSizes } from '@material-ui/core/styles';
+import { ThemeProvider } from '@material-ui/styles';
+import Typography from '@material-ui/core/Typography';
+
+let theme = createMuiTheme();
+theme = responsiveFontSizes(theme);
+
+export default function ResponsiveFontSizes() {
+  return (
+    <ThemeProvider theme={theme}>
+      <Typography variant="h3">Responsive h3</Typography>
+    </ThemeProvider>
+  );
+}
