+++ docs/webpackBaseConfig.js
@@ -14,7 +14,6 @@ module.exports = {
       '@material-ui/styles': path.resolve(__dirname, '../packages/material-ui-styles/src'),
       '@material-ui/utils': path.resolve(__dirname, '../packages/material-ui-utils/src'),
       '@material-ui/system': path.resolve(__dirname, '../packages/material-ui-system/src'),
-      '@material-ui/css-utils': path.resolve(__dirname, '../packages/material-ui-css-utils/src'),
       docs: path.resolve(__dirname, '../docs'),
     },
   },
