--- packages/material-ui-css-utils/docs/README.hbs
@@ -1,19 +1,0 @@
-# @material-ui/css-utils
-
-The `@material-ui/css-utils` package provides helper functions that you can use to manipulate CSS-in-JS values.
-
-## Installation
-
-Install the package in your project directory with:
-
-```sh
-// with npm
-npm install @material-ui/css-utils
-
-// with yarn
-yarn add @material-ui/css-utils
-```
-
-## API
-
-{{>main}}
