--- packages/material-ui-css-utils/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@material-ui/css-utils",
-  "private": false,
-  "author": "Material-UI Team",
-  "version": "4.0.0-alpha.0",
-  "description": "Material-UI CSS Kit - CSS helpers for Material-UI.",
-  "main": "./src/index.js",
-  "keywords": [
-    "material design",
-    "material-ui",
-    "css",
-    "responsiveness",
-    "typography"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mui-org/material-ui.git"
-  },
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mui-org/material-ui/issues"
-  },
-  "homepage": "https://github.com/mui-org/material-ui/tree/next/packages/material-ui-css-utils",
-  "scripts": {
-    "build": "yarn build:cjs && yarn build:esm && yarn build:es && yarn build:copy-files",
-    "build:cjs": "cross-env NODE_ENV=production BABEL_ENV=cjs babel --config-file ../../babel.config.js ./src --out-dir ./build --ignore \"**/*.test.js\"",
-    "build:esm": "cross-env NODE_ENV=production BABEL_ENV=esm babel --config-file ../../babel.config.js ./src --out-dir ./build/esm --ignore \"**/*.test.js\"",
-    "build:es": "cross-env NODE_ENV=production BABEL_ENV=es babel --config-file ../../babel.config.js ./src --out-dir ./build/es --ignore \"**/*.test.js\"",
-    "build:copy-files": "node ../../scripts/copy-files.js",
-    "prebuild": "rimraf build",
-    "release": "yarn build && npm publish build --tag next",
-    "test": "cd ../../ && cross-env NODE_ENV=test mocha 'packages/material-ui-css-utils/**/*.test.js' --exclude '**/node_modules/**'",
-    "docs": "jsdoc2md --template docs/README.hbs src/*.js > README.md"
-  },
-  "peerDependencies": {},
-  "dependencies": {
-    "@babel/runtime": "^7.2.0"
-  },
-  "devDependencies": {
-    "jsdoc-to-markdown": "^4.0.1"
-  },
-  "sideEffects": false,
-  "publishConfig": {
-    "access": "public"
-  },
-  "engines": {
-    "node": ">=8.0.0"
-  }
-}
