--- packages/material-ui-css-utils/src/alignProperty.js
@@ -1,8 +1,0 @@
-function alignProperty({ size, grid }) {
-  const sizeBelow = size - (size % grid);
-  const sizeAbove = sizeBelow + grid;
-
-  return size - sizeBelow < sizeAbove - size ? sizeBelow : sizeAbove;
-}
-
-export default alignProperty;
