--- packages/material-ui-css-utils/src/index.js
@@ -1,3 +1,0 @@
-export { default as responsiveProperty } from './responsiveProperty';
-export { default as alignProperty } from './alignProperty';
-export { default as fontGrid } from './fontGrid';
