+++ packages/material-ui/package.json
@@ -40,7 +40,6 @@
     "@material-ui/styles": "^4.0.0-beta.0",
     "@material-ui/system": "^4.0.0-beta.0",
     "@material-ui/utils": "^4.0.0-beta.0",
-    "@material-ui/css-utils": "^4.0.0-alpha.0",
     "@types/react-transition-group": "^2.0.16",
     "clsx": "^1.0.2",
     "csstype": "^2.5.2",
