+++ packages/material-ui/src/styles/index.js
@@ -3,9 +3,9 @@ export { default as createMuiTheme } from './createMuiTheme';
 export { default as createStyles } from './createStyles';
 export { default as makeStyles } from './makeStyles';
 export { default as MuiThemeProvider } from './MuiThemeProvider';
+export { default as responsiveFontSizes } from './responsiveFontSizes';
 export { default as styled } from './styled';
 export * from './transitions';
 export { default as useTheme } from './useTheme';
 export { default as withStyles } from './withStyles';
 export { default as withTheme } from './withTheme';
-export { default as responsiveTypography } from './responsiveTypography';
