+++ scripts/sizeSnapshot/webpack.config.js
@@ -135,7 +135,6 @@ module.exports = async function webpackConfig() {
         '@material-ui/styles': path.join(workspaceRoot, 'packages/material-ui-styles/build'),
         '@material-ui/system': path.join(workspaceRoot, 'packages/material-ui-system/build'),
         '@material-ui/utils': path.join(workspaceRoot, 'packages/material-ui-utils/build'),
-        '@material-ui/css-utils': path.join(workspaceRoot, 'packages/material-ui-css-utils/build'),
       },
     },
   };
